pyproject.toml
[tool.poetry]
name = "challenge-back-end"
version = "0.1.0"
description = "Gerenciamento de dependencias"
authors = ["Jefferson Fragoso de Souza <jeffersonfragoso@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
django = "^4.2.5"
drf-yasg = "^1.21.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
bandit = "^1.7.5"
safety = "^2.3.5"
pre-commit = "^3.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
